Remove now-unnecesary Registry values
The values deleted in this commit instruct Windows to display numerous
metadata items in the Properties > Details window. However, because my
filter does not actually extract any of these properties, all the values
displayed are blank. In addition, I no longer need to see the character
count property now that the IFilter extracts the text correctly.